python-core/requirements.txt
# Smart Ticket System - Python Core Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# AI Framework
langchain>=0.1.0
langchain-openai>=0.0.5

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client (for external API calls)
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0